[chaos] Move header and footer
