--- src/header/index.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-import { Bar } from 'header/bar'
-import { MobileOptions } from 'header/mobile-options'
-
-export const Header = () => <>
-  <Bar />
-  <MobileOptions />
-</>
