+++ src/layout/footer/nav-buttons/index.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { makeStyles, Box, Grid } from '@material-ui/core'
 
-import { About } from 'footer/nav-buttons/about'
-import { Donate } from 'footer/nav-buttons/donate'
-import { PrivacyPolicy } from 'footer/nav-buttons/privacy-policy'
-import { Terms } from 'footer/nav-buttons/terms'
+import { About } from 'layout/footer/nav-buttons/about'
+import { Donate } from 'layout/footer/nav-buttons/donate'
+import { PrivacyPolicy } from 'layout/footer/nav-buttons/privacy-policy'
+import { Terms } from 'layout/footer/nav-buttons/terms'
 
 const useStyles = makeStyles(theme => ({
   navButtons: {
