+++ src/layout/header/desktop-buttons/contact.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Grid } from '@material-ui/core'
 
-import { Link } from 'header/link'
-import { actions } from 'header/constants'
+import { Link } from 'layout/header/link'
+import { actions } from 'layout/header/constants'
 
 export const Contact = () => <Grid item md={3} lg={3} xl={3}>
   <Link>{actions.contact}</Link>
