+++ src/layout/header/icon.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core'
 
-import Source from 'header/icon.png'
+import Source from 'layout/header/icon.png'
 
 const useStyles = makeStyles({ root: { height: 50 } })
