+++ src/layout/header/icon.png
