+++ src/layout/header/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import { Bar } from 'layout/header/bar'
+import { MobileOptions } from 'layout/header/mobile-options'
+
+export const Header = () => <>
+  <Bar />
+  <MobileOptions />
+</>
