+++ src/layout/header/mobile-buttons/hooks/use-toggle-callbacks.js
@@ -1,7 +1,7 @@
 import { useCallback } from 'react'
 import { useDispatch } from 'react-redux'
 
-import { actions } from 'header/slices/mobile'
+import { actions } from 'layout/header/slices/mobile'
 
 export const useToggleCallbacks = () => {
   const dispatch = useDispatch()
