+++ src/layout/header/mobile-options/about.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ListItem, ListItemText } from '@material-ui/core'
 
-import { actions } from 'header/constants'
+import { actions } from 'layout/header/constants'
 
 export const About = () => <ListItem button>
   <ListItemText>{actions.about}</ListItemText>
