+++ src/layout/header/mobile-options/hooks/use-visibility.js
