+++ src/layout/header/slices/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from '@reduxjs/toolkit'
 
-import { mobile } from 'header/slices/mobile'
+import { mobile } from 'layout/header/slices/mobile'
 
 export const header = combineReducers({ mobile: mobile.reducer })
