+++ src/startup/reducers.js
@@ -1,6 +1,6 @@
 import { combineReducers } from '@reduxjs/toolkit'
 
-import { header } from 'header/slices'
 import { course } from 'course/slices'
+import { header } from 'layout/header/slices'
 
 export const reducers = combineReducers({ header, course })
